fix: clear mechanical lint findings and browsergate discovery projection
golangci: lowercase ST1005 error strings, %w errorlint wrapping,
errors.As conversions, De Morgan and tagged-switch simplifications,
dead signalAll removal, and a func-main unused exclusion for the
harness-built helper commands.

windowsjob: move windows-only start-gate publication, writeAll, and
ownerReadyByte into windows-tagged files (mirroring start_gate_linux.go);
settledInputOutcome is a test helper and moves into the windows test.

browsergate: playwrightDiscoveryEnvironment detects Windows-cased
canonical spellings (Path, SystemRoot) and projects them
case-insensitively on any host, fixing the linux contract run.

+++ internal/processowner/linuxsubreaper/executable_linux.go
@@ -207,7 +207,7 @@ func (authority *executableAuthority) startEvidence(
 		Platform:        ownerprotocol.PlatformLinuxSubreaper,
 		ProcessID:       root.PID,
 		ProcessInstance: strconv.FormatUint(root.StartTimeTicks, 10),
-		Executable:      ownerprotocol.NewObjectIdentity64(uint64(metadata.Dev), metadata.Ino),
+		Executable:      ownerprotocol.NewObjectIdentity64(metadata.Dev, metadata.Ino),
 	}, nil
 }
